testapp/requirements.txt
Django>=2.0.13
python-dateutil
